Add files via upload
Tasker is a simple appointments application written in Bash.
Currently status:
It will display appointments(tasks) scheduled for the current month and next month. Data is held in a simple text file.
To be added:
Options are in hand to be able to add, edit and delete tasks. Newly added tasks will populate the data file depending on whether they are one-offs or occurring daily, weekly, monthly or annually, with recurrences set for a limited or infinite number of instances.